Refactor to reenforce Task pattern
SkraprWorker now executes Tasks and thus tasks can be added to the main flow by sub tasks.

Rules are now generalized and multiple types of rules can be defined. Rules can examine the SkraprFrameState to determine if the rule matches. Currently only the UriPatternRule exists.

General refinements to add comments, favor ILists when the intent is a indexed list where order matters.

Add ability for tasks to be marked as disabled.

Main process now exits when all tasks on the main flow are completed.